feat(EntityFramework): Extend MasaDbContextOptionsBuilder
