+++ Masa.Contrib.sln
@@ -316,7 +316,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Masa.BuildingBlocks.BasicAb
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Masa.Contrib.BasicAbility.Mc.Tests", "test\Masa.Contrib.BasicAbility.Mc.Tests\Masa.Contrib.BasicAbility.Mc.Tests.csproj", "{23633E49-F11A-4D14-899A-E2599C8182CE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Masa.Contrib.Data.UoW.EF", "src\Data\Masa.Contrib.Data.UoW.EF\Masa.Contrib.Data.UoW.EF.csproj", "{FC4E526A-DBFC-406A-8ED3-64983B67F688}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Masa.Contrib.Data.UoW.EF", "src\Data\Masa.Contrib.Data.UoW.EF\Masa.Contrib.Data.UoW.EF.csproj", "{FC4E526A-DBFC-406A-8ED3-64983B67F688}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -1294,14 +1294,6 @@ Global
 		{4E237346-F948-46AC-801B-492545978280}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4E237346-F948-46AC-801B-492545978280}.Release|x64.ActiveCfg = Release|Any CPU
 		{4E237346-F948-46AC-801B-492545978280}.Release|x64.Build.0 = Release|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|x64.Build.0 = Debug|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|x64.ActiveCfg = Release|Any CPU
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|x64.Build.0 = Release|Any CPU
 		{75A25CF6-9BA4-46F5-8BC3-90396230CB64}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{75A25CF6-9BA4-46F5-8BC3-90396230CB64}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{75A25CF6-9BA4-46F5-8BC3-90396230CB64}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1326,6 +1318,14 @@ Global
 		{23633E49-F11A-4D14-899A-E2599C8182CE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{23633E49-F11A-4D14-899A-E2599C8182CE}.Release|x64.ActiveCfg = Release|Any CPU
 		{23633E49-F11A-4D14-899A-E2599C8182CE}.Release|x64.Build.0 = Release|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Debug|x64.Build.0 = Debug|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|x64.ActiveCfg = Release|Any CPU
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1479,11 +1479,11 @@ Global
 		{8D84666E-C79E-4D49-B73D-360E62D312EF} = {5DFAF4A2-ECB5-46E4-904D-1EA5F48B2D48}
 		{7C4B88FA-3681-4C29-BC3F-0FDB111C5681} = {FBD326D3-E59C-433E-A88E-14E179E3093D}
 		{592297DE-DA72-452D-9D88-61EE882FE9A6} = {38E6C400-90C0-493E-9266-C1602E229F1B}
-		{FC4E526A-DBFC-406A-8ED3-64983B67F688} = {E33ADF54-4D35-49B7-BDA6-412587CA39FF}
 		{4E237346-F948-46AC-801B-492545978280} = {38E6C400-90C0-493E-9266-C1602E229F1B}
 		{75A25CF6-9BA4-46F5-8BC3-90396230CB64} = {5DFAF4A2-ECB5-46E4-904D-1EA5F48B2D48}
 		{DA816A33-F164-4456-92DD-A672BAD1A6B1} = {0D34A7F0-DC77-4789-A136-93089CBD15C3}
 		{23633E49-F11A-4D14-899A-E2599C8182CE} = {38E6C400-90C0-493E-9266-C1602E229F1B}
+		{FC4E526A-DBFC-406A-8ED3-64983B67F688} = {E33ADF54-4D35-49B7-BDA6-412587CA39FF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {40383055-CC50-4600-AD9A-53C14F620D03}
