+++ test/Masa.Contrib.BasicAbility.Mc.Tests/_Imports.cs
@@ -1,14 +1,13 @@
 // Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using System.Threading.Tasks;
+global using Masa.BuildingBlocks.BasicAbility.Mc;
 global using Masa.BuildingBlocks.BasicAbility.Mc.Model;
+global using Masa.Contrib.BasicAbility.Mc.Service;
 global using Masa.Utils.Caller.Core;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Moq;
 global using System;
-global using Masa.Contrib.BasicAbility.Mc.Service;
-global using Masa.BuildingBlocks.Ddd.Domain.Repositories;
-global using Masa.BuildingBlocks.BasicAbility.Mc;
-global using Microsoft.Extensions.DependencyInjection;
 global using System.Collections.Generic;
+global using System.Threading.Tasks;
