+++ test/Masa.Contrib.Data.EntityFrameworkCore.Tests/_Imports.cs
@@ -6,7 +6,6 @@ global using Masa.BuildingBlocks.Data.Contracts.DataFiltering;
 global using Masa.BuildingBlocks.Data.Contracts.Paginated;
 global using Masa.Contrib.Data.Contracts.EF;
 global using Masa.Contrib.Data.EntityFrameworkCore.Filters;
-global using Masa.Contrib.Data.EntityFrameworkCore.Sqlite;
 global using Masa.Contrib.Data.EntityFrameworkCore.Tests.Internal;
 global using Masa.Contrib.Data.EntityFrameworkCore.Tests.Models;
 global using Microsoft.EntityFrameworkCore;
