+++ test/Masa.Contrib.Data.UoW.EF.Tests/_Imports.cs
@@ -7,7 +7,6 @@ global using Masa.BuildingBlocks.Data.Options;
 global using Masa.BuildingBlocks.Data.UoW;
 global using Masa.BuildingBlocks.Dispatcher.Events;
 global using Masa.Contrib.Data.EntityFrameworkCore;
-global using Masa.Contrib.Data.EntityFrameworkCore.Sqlite;
 global using Microsoft.AspNetCore.Builder;
 global using Microsoft.Data.Sqlite;
 global using Microsoft.EntityFrameworkCore;
