+++ test/Masa.Contrib.Ddd.Domain.Repository.EF.Tests/_Imports.cs
@@ -8,7 +8,6 @@ global using Masa.BuildingBlocks.Ddd.Domain.Repositories;
 global using Masa.BuildingBlocks.Ddd.Domain.Values;
 global using Masa.BuildingBlocks.Dispatcher.Events;
 global using Masa.Contrib.Data.EntityFrameworkCore;
-global using Masa.Contrib.Data.EntityFrameworkCore.Sqlite;
 global using Masa.Contrib.Data.UoW.EF;
 global using Masa.Contrib.Ddd.Domain.Repository.EF.CustomRepository.Tests.Repositories;
 global using Masa.Contrib.Ddd.Domain.Repository.EF.Tests.Domain.Entities;
@@ -16,7 +15,6 @@ global using Masa.Contrib.Ddd.Domain.Repository.EF.Tests.Domain.Repositories;
 global using Masa.Contrib.Ddd.Domain.Repository.EF.Tests.Infrastructure;
 global using Microsoft.Data.Sqlite;
 global using Microsoft.EntityFrameworkCore;
-global using Microsoft.EntityFrameworkCore.Storage;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Moq;
