+++ test/Masa.Contrib.Dispatcher.IntegrationEvents.Dapr.Tests/_Imports.cs
@@ -9,11 +9,10 @@ global using Masa.BuildingBlocks.Dispatcher.IntegrationEvents;
 global using Masa.BuildingBlocks.Dispatcher.IntegrationEvents.Logs;
 global using Masa.Contrib.Dispatcher.IntegrationEvents.Dapr.Options;
 global using Masa.Contrib.Dispatcher.IntegrationEvents.Processor;
+global using Masa.Contrib.Dispatcher.IntegrationEvents.Tests.Infrastructure;
 global using Masa.Utils.Models.Config;
 global using Microsoft.Extensions.DependencyInjection;
-global using Microsoft.Extensions.Logging.Abstractions;
 global using Microsoft.Extensions.Options;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Moq;
 global using System.Reflection;
-global using Masa.Contrib.Dispatcher.IntegrationEvents.Tests.Infrastructure;
