+++ test/Masa.Contrib.Identity.IdentityModel.Tests/_Imports.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
 global using Masa.BuildingBlocks.Identity.IdentityModel;
-global using Masa.Contrib.Identity.IdentityModel;
 global using Masa.Contrib.Identity.IdentityModel.Const;
 global using Microsoft.AspNetCore.Http;
 global using Microsoft.Extensions.DependencyInjection;
