+++ test/Masa.Contrib.Isolation.UoW.EF.Tests/_Imports.cs
@@ -8,10 +8,8 @@ global using Masa.BuildingBlocks.Dispatcher.Events;
 global using Masa.BuildingBlocks.Isolation.Environment;
 global using Masa.BuildingBlocks.Isolation.MultiTenant;
 global using Masa.BuildingBlocks.Isolation.Options;
-global using Masa.Contrib.Configuration;
 global using Masa.Contrib.Data.Contracts.EF;
 global using Masa.Contrib.Data.EntityFrameworkCore;
-global using Masa.Contrib.Data.EntityFrameworkCore.Sqlite;
 global using Masa.Contrib.Isolation.MultiEnvironment;
 global using Masa.Contrib.Isolation.MultiTenant;
 global using Microsoft.AspNetCore.Builder;
