+++ test/Masa.Contrib.SearchEngine.AutoComplete.ElasticSearch.Tests/_Imports.cs
@@ -4,7 +4,6 @@
 global using Masa.BuildingBlocks.SearchEngine.AutoComplete;
 global using Masa.BuildingBlocks.SearchEngine.AutoComplete.Options;
 global using Masa.Contrib.SearchEngine.AutoComplete.ElasticSearch.Tests.Model;
-global using Masa.Utils.Data.Elasticsearch;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Nest;
